Refactor request e response
