+++ src/main/java/com/stock/flex/config/TestConfig.java
@@ -1,6 +1,6 @@
 package com.stock.flex.config;
 
-import com.stock.flex.security.services.DatabaseService;
+import com.stock.flex.security.service.DatabaseService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
