+++ src/main/java/com/stock/flex/security/service/DatabaseService.java
@@ -1,4 +1,4 @@
-package com.stock.flex.security.services;
+package com.stock.flex.security.service;
 
 import com.stock.flex.entity.Person;
 import com.stock.flex.entity.enums.Role;
@@ -21,16 +21,13 @@ public class DatabaseService {
 
 		System.out.println("Initializing database...");
 		
-		final Person user1 = new Person("Emma", "emma@mail.com", passwordEncoder.encode("111"));
-		final Person user2 = new Person("Jhon", "jhon@mail.com", passwordEncoder.encode("222"));
-		final Person admin = new Person("Anna", "anna@mail.com", passwordEncoder.encode("333"));
+		final Person user = new Person("Emma", "emma@mail.com", passwordEncoder.encode("111"));
+		final Person admin = new Person("Anna", "anna@mail.com", passwordEncoder.encode("222"));
 		
 		admin.addRole(Role.ADMIN);
 	
-		System.out.println(personService.create(user1));
-		System.out.println(personService.create(user2));
+		System.out.println(personService.create(user));
 		System.out.println(personService.create(admin));
 		
-		System.out.println("Database initialized!");
 	}
 }
