+++ src/main/java/com/stock/flex/security/service/UserDetailsServiceImpl.java
@@ -1,4 +1,4 @@
-package com.stock.flex.security.services;
+package com.stock.flex.security.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
