+++ src/main/resources/application.properties
@@ -1,16 +1,22 @@
-spring.h2.console.enabled = true
+# Configurações do H2
+spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
 
-spring.datasource.url=jdbc:h2:mem:testdb
+# Configurações do banco de dados H2 com a porta 8081
+spring.datasource.url=jdbc:h2:mem:testdb;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE;DATABASE_TO_UPPER=false;INIT=CREATE SCHEMA IF NOT EXISTS PUBLIC
 spring.datasource.username=sa
 spring.datasource.password=
 spring.datasource.driver-class-name=org.h2.Driver
 
+# Configurações do JPA
 spring.jpa.show-sql=false
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.properties.jakarta.persistence.sharedCache.mode=UNSPECIFIED
 spring.jpa.open-in-view=false
 
+# Configurações do JWT
 jwt.secret.key=aLongSecretStringWhoseBitnessIsEqualToOrGreaterThanTheBitnessOfTheTokenEncryptionAlgorithm
 jwt.expiration=86400000
 
+# Configuração da porta 8081
+server.port=8081
